refactor: refactor scripts
- consolidate types scripts into one file
- rename ci:download-vscode to download-vscode for consistency
- alphabetically order scripts in package.json
