+++ scripts/download-vscode.js
